+++ api/src/main/java/org/allaymc/api/plugin/PluginDescriptor.java
@@ -12,61 +12,81 @@ import java.util.List;
  */
 public interface PluginDescriptor {
 
+    /**
+     * Checks if a descriptor is valid.
+     *
+     * @param descriptor the descriptor to check
+     */
+    @SuppressWarnings("DataFlowIssue")
     static void checkDescriptorValid(PluginDescriptor descriptor) {
         Preconditions.checkNotNull(descriptor.getName(), "Plugin name cannot be null");
         Preconditions.checkNotNull(descriptor.getEntrance(), "Plugin entrance cannot be null");
         Preconditions.checkNotNull(descriptor.getVersion(), "Plugin version cannot be null");
         Preconditions.checkNotNull(descriptor.getAuthors(), "Plugin authors cannot be null");
-        // noinspection DataFlowIssue
+        Preconditions.checkNotNull(descriptor.getDescription(), "Plugin description cannot be null");
+        Preconditions.checkNotNull(descriptor.getAPIVersion(), "Plugin api version cannot be null");
+        Preconditions.checkNotNull(descriptor.getDependencies(), "Plugin dependencies cannot be null");
+        for (var dependency : descriptor.getDependencies()) {
+            Preconditions.checkNotNull(dependency.name(), "Dependency name cannot be null");
+            Preconditions.checkArgument(!dependency.name().isBlank(), "Dependency name cannot be blank");
+        }
+        Preconditions.checkNotNull(descriptor.getWebsite(), "Plugin website cannot be null");
         Preconditions.checkNotNull(Semver.coerce(descriptor.getVersion()), "Plugin version cannot be coerced (Use https://semver.org/)");
     }
 
     /**
-     * Get the name of the plugin.
+     * Gets the name of the plugin.
      *
      * @return the name of the plugin
      */
     String getName();
 
     /**
-     * Get the entrance of the plugin.
+     * Gets the entrance of the plugin.
      *
      * @return the entrance of the plugin
      */
     String getEntrance();
 
     /**
-     * Get the description of the plugin.
-     *
-     * @return the description of the plugin
-     */
-    String getDescription();
-
-    /**
-     * Get the version of the plugin.
+     * Gets the version of the plugin.
      *
      * @return the version of the plugin
      */
     String getVersion();
 
     /**
-     * Get the authors of the plugin.
+     * Gets the authors of the plugin.
      *
      * @return the authors of the plugin
      */
     List<String> getAuthors();
 
     /**
-     * Get the dependencies of the plugin.
+     * Gets the description of the plugin.
      *
-     * @return the dependencies of the plugin
+     * @return the description of the plugin, or a blank string if is not specified
+     */
+    String getDescription();
+
+    /**
+     * Gets the api version of the plugin.
+     *
+     * @return the api version of the plugin, or a blank string if is not specified
+     */
+    String getAPIVersion();
+
+    /**
+     * Gets the dependencies of the plugin.
+     *
+     * @return the dependencies of the plugin, or an empty list if is not specified
      */
     List<PluginDependency> getDependencies();
 
     /**
-     * Get the website of the plugin.
+     * Gets the website of the plugin.
      *
-     * @return the website of the plugin
+     * @return the website of the plugin, or a blank string if is not specified
      */
     String getWebsite();
 }