+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Unless otherwise specified, any version comparison below is the comparison of se
 - (API) Introduced option `ServerSettings.NetworkSettings.enablev6` to close ipv6 if needed.
 - (API) Introduced field `PlayerRespawnEvent.respawnLocation` which can be changed to modify the respawn location of a player after death.
 - (API) Introduced field `LiquidHardenEvent.hardenedBlockPosition` which is the position of the hardened block.
+- (API) Introduced `api_version` for plugin descriptor, which can set the api version requirement of a plugin.
 - Introduced dirty flag for block layers in chunk section. Now blocks will only be rewritten to the database if they are changed. This
   would speed up the time used during server shutdown significantly if there are many only loaded chunks.
 - Added support for basic multi-version. The server now support 1.21.80 - 1.21.100 client to join.